Refactor the 'render' command as mentioned in #4
