+++ astraledit-bukkit-plugin/src/main/java/com/github/shynixn/astraledit/bukkit/logic/business/SelectionCommandExecutor.java
@@ -6,6 +6,7 @@ import com.github.shynixn.astraledit.api.bukkit.business.entity.Selection;
 import com.github.shynixn.astraledit.bukkit.AstralEditPlugin;
 import com.github.shynixn.astraledit.bukkit.Permission;
 import com.github.shynixn.astraledit.bukkit.logic.business.command.AutoRotateCommand;
+import com.github.shynixn.astraledit.bukkit.logic.business.command.RenderCommand;
 import com.github.shynixn.astraledit.bukkit.logic.lib.SimpleCommandExecutor;
 import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
@@ -31,6 +32,7 @@ class SelectionCommandExecutor extends SimpleCommandExecutor.Registered {
         this.manager = manager;
 
         this.commands.add(new AutoRotateCommand(manager));
+        this.commands.add(new RenderCommand(plugin));
     }
 
     /**
@@ -49,9 +51,7 @@ class SelectionCommandExecutor extends SimpleCommandExecutor.Registered {
             }
         }
 
-        if (args.length == 1 && args[0].equalsIgnoreCase("render") && Permission.RENDER.hasPermission(player))
-            this.createRenderCommand(player);
-        else if (args.length == 1 && args[0].equalsIgnoreCase("join") && Permission.JOIN.hasPermission(player))
+        if (args.length == 1 && args[0].equalsIgnoreCase("join") && Permission.JOIN.hasPermission(player))
             this.combineCommand(player);
         else if (args.length == 1 && args[0].equalsIgnoreCase("tear") && Permission.TEAR.hasPermission(player))
             this.unCombineCommand(player);
@@ -455,23 +455,6 @@ class SelectionCommandExecutor extends SimpleCommandExecutor.Registered {
         });
     }
 
-    /**
-     * Creates a new rendered object for the player
-     *
-     * @param player player
-     */
-    private void createRenderCommand(final Player player) {
-        this.runAsyncTask(() -> {
-            player.sendMessage(AstralEditPlugin.PREFIX_SUCCESS + "Rendering WorldEdit-Selection asynchronously...");
-            final Selection selection = AstralEditApi.INSTANCE.render(player);
-            if (selection == null) {
-                player.sendMessage(AstralEditPlugin.PREFIX_ERROR + "Failed rendering WE selection!");
-                player.sendMessage(AstralEditPlugin.PREFIX_ERROR + "Check if you selected an area with Worldedit.");
-            } else {
-                player.sendMessage(AstralEditPlugin.PREFIX_SUCCESS + "Finished rendering selection.");
-            }
-        });
-    }
 
     /**
      * Runs task asynchronously
